refactor(boundaries): break onboarding ↔ onboarding_tui circular import (refs #1367)
Move ``run_onboarding`` from ``pollypm.onboarding`` into
``pollypm.onboarding_tui``. ``onboarding.py`` no longer references
``onboarding_tui`` at all (the function-level lazy
``from pollypm.onboarding_tui import run_onboarding_app`` is gone), so
the cycle collapses to a one-way ``onboarding_tui -> onboarding`` edge.

Also promote the previously-lazy ``seed_demo_project_task`` import in
``onboarding_tui`` to the module-level block now that the cycle is gone.

``run_onboarding`` accesses ``_launch_onboarding_experience`` and
``_render_welcome_back_summary`` through the ``pollypm.onboarding``
module object so the existing test monkeypatch seam
(``monkeypatch.setattr("pollypm.onboarding._launch_onboarding_experience", ...)``)
keeps working unchanged.

CLI lazy import sites and the one test that called the entry point via
``pollypm.onboarding.run_onboarding`` are updated to
``pollypm.onboarding_tui.run_onboarding``.
